Added move logic and refactoring code

+++ src/main/java/com/wallapop/business/Direction.java
@@ -6,6 +6,6 @@ interface Direction {
     Character getCharacter();
     Direction turnLeft();
     Direction turnRight();
-    Coordinate goForward(Integer positionX, Integer positionY);
-    Coordinate goBackward(Integer positionX, Integer positionY);
+    Coordinate goForward(Long positionX, Long positionY);
+    Coordinate goBackward(Long positionX, Long positionY);
 }
