+++ src/main/java/com/wallapop/business/EastDirection.java
@@ -14,12 +14,12 @@ public class EastDirection implements Direction {
     }
 
     @Override
-    public Coordinate goForward(Integer positionX, Integer positionY) {
+    public Coordinate goForward(Long positionX, Long positionY) {
         return new Coordinate(positionX + 1, positionY);
     }
 
     @Override
-    public Coordinate goBackward(Integer positionX, Integer positionY) {
+    public Coordinate goBackward(Long positionX, Long positionY) {
         return new Coordinate(positionX - 1, positionY);
     }
